internal/forms/forms.go
package forms

import (
	"fmt"
	"net/url"
	"strings"

	"github.com/asaskevich/govalidator"
)

// Form creates a custom form struct and embeds a url.Values object
type Form struct {
	url.Values
	Errors errors
}

// Valid returns true if there are no errors, otherwise false
func (f *Form) Valid() bool {
	return len(f.Errors) == 0
}

// New initializes a form struct
func New(data url.Values) *Form {
	return &Form{
		data,
		errors(map[string][]string{}),
	}
}
// Required checks if form field is in post and not empty
func (f *Form) Required(fields ...string){
	for _,field := range fields {
		value := f.Get(field) 
		if strings.TrimSpace(value) == "" {
			f.Errors.Add(field,"This field cannot be empty")
		}
	}
}

// Has checks if form field is in post and not empty
func (f *Form) Has(field string ) bool {
	x := f.Get(field)
	if x == "" {
		return false
	}
	return true
}
//MinLenght check for string minimum strenght
func (f *Form) MinLenght(field string, lenght int)  {
	x := f.Get(field)
	if len(x) < lenght {
		f.Errors.Add(field,fmt.Sprintf("This field must be at least %d characters long",lenght ))
		
	}
}

//IsEmail Check for valid email addres
func (f *Form) IsEmail(field string)  {
	if !govalidator.IsEmail(f.Get(field)) {
		f.Errors.Add(field,"Invalid Email Addres")
	}
}